requirements.txt
fastapi
uvicorn
python-dotenv
openai
supabase-py
jinja2
python-multipart
